apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
PyYAML==6.0.2
httpx==0.27.2
python-dotenv==1.0.1
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
